// retailer/src/app/cart/model/delivery.ts
export class Delivery {
    name: string;
    lastName: string;
    identityNumber: string;
    phone: string;
    email: string;
    addressOne: string;
    taxPlace: string;
    taxNo: string;
    state: string;
    city: string;
    zipCode: string;
    country: string;
    countryCode: string;
}